allow async methods to return a value to python client
Before, `SyftController.ProcessMessage` was synchronous and was required
to return a string value.

I have changed the function definition to return void, and it is passed
a callback which you should call with a string, what you pass the
callback is passed back to the client.

This allows you to leave the main thread and call the callback at a
later date.

Grid takes advantage of this, when you call `getResults`, they may not
be available yet so we wait for them.  The only wait to wait and check
again is to pause, since this was running on the main thread it would
completely freeze unity, not ideal!

check out `Grid.cs` for an example of how I used this.